+++ srcjs/__tests__/utils.test.js
@@ -1,4 +1,12 @@
-import { classNames, getFirstDefined, getStrIncludesLocale, strIncludes, get, set } from '../utils'
+import {
+  classNames,
+  getFirstDefined,
+  getStrIncludesLocale,
+  strIncludes,
+  get,
+  set,
+  invertObj
+} from '../utils'
 
 test('classNames', () => {
   expect(classNames('')).toEqual('')
@@ -70,4 +78,11 @@ test('set', () => {
   const obj = { 1: 2 }
   set(obj, [1], 5)
   expect(obj[1]).toEqual(2)
+})
+
+test('invertObj', () => {
+  expect(invertObj({})).toEqual({})
+  expect(invertObj({ abc: 0 })).toEqual({ 0: 'abc' })
+  expect(invertObj({ a: [1, 2, 3] })).toEqual({ 1: 'a', 2: 'a', 3: 'a' })
+  expect(invertObj({ abc: 0, a: [1, 2, 3], b: [3, 0] })).toEqual({ 0: 'b', 1: 'a', 2: 'a', 3: 'b' })
 })